move EmitterBuilder to main package
