+++ outwatch/src/main/scala/outwatch/dom/package.scala
@@ -4,6 +4,7 @@ import com.raquo.domtypes.generic.keys
 import outwatch.dom.helpers.BasicStyleBuilder
 
 package object dom extends ManagedSubscriptions {
+  type EmitterBuilder[+O, +R] = EmitterBuilderExecution[O, R, EmitterBuilder.Execution]
 
   //TODO: invent typeclass CanBuildStyle[F[_]]
   @inline implicit def StyleIsBuilder[T](style: keys.Style[T]): BasicStyleBuilder[T] = new BasicStyleBuilder[T](style.cssName)
